API: move all wlroots includes into their own header
There are three reasons to do this:

- Having #define and #undef everywhere a wlroots header is included is
not nice.
- It is easy to forget to include `wlr/config.h` or `wayfire/config.h`,
which are needed for WF_HAS_WAYLAND/WLR_HAS_X11_BACKEND.
- `extern "C"` seems to not play nicely with clangd, making code
completions slow in files where it is present.
