+++ src/api/wayfire/core.hpp
@@ -10,38 +10,8 @@
 #include <wayfire/nonstd/observer_ptr.h>
 #include <wayfire/config/config-manager.hpp>
 
-extern "C"
-{
-    struct wlr_backend;
-    struct wlr_renderer;
-    struct wlr_seat;
-    struct wlr_cursor;
-    struct wlr_data_device_manager;
-    struct wlr_data_control_manager_v1;
-    struct wlr_gamma_control_manager_v1;
-    struct wlr_xdg_output_manager_v1;
-    struct wlr_export_dmabuf_manager_v1;
-    struct wlr_server_decoration_manager;
-    struct wlr_xdg_decoration_manager_v1;
-    struct wlr_input_inhibit_manager;
-    struct wlr_virtual_keyboard_manager_v1;
-    struct wlr_virtual_pointer_manager_v1;
-    struct wlr_idle;
-    struct wlr_idle_inhibit_manager_v1;
-    struct wlr_screencopy_manager_v1;
-    struct wlr_foreign_toplevel_manager_v1;
-    struct wlr_pointer_gestures_v1;
-    struct wlr_relative_pointer_manager_v1;
-    struct wlr_pointer_constraints_v1;
-    struct wlr_tablet_manager_v2;
-    struct wlr_input_method_manager_v2;
-    struct wlr_text_input_manager_v3;
-    struct wlr_presentation;
-    struct wlr_gtk_primary_selection_device_manager;
-    struct wlr_primary_selection_v1_device_manager;
-
 #include <wayland-server.h>
-}
+#include <wayfire/nonstd/wlroots.hpp>
 
 namespace wf
 {
