+++ src/output/gtk-shell.cpp
@@ -1,7 +1,3 @@
-extern "C"
-{
-#include <wlr/types/wlr_surface.h>
-}
 #include "gtk-shell.hpp"
 #include "gtk-shell-protocol.h"
 #include <wayfire/util/log.hpp>
@@ -11,6 +7,7 @@ extern "C"
 #include <map>
 #include <wayfire/signal-definitions.hpp>
 #include <wayfire/gtk-shell.hpp>
+#include <wayfire/nonstd/wlroots-full.hpp>
 #define GTK_SHELL_VERSION 3
 
 struct wf_gtk_shell
