+++ uncrustify.ini
@@ -1061,6 +1061,7 @@ mod_enum_last_comma             = force   # ignore/add/remove/force
 # Add or remove indentation of preprocessor directives inside #if blocks
 # at brace level 0 (file-level).
 pp_indent                       = add   # ignore/add/remove/force
+pp_indent_count                 = 4
 
 # Whether to indent the code between #if, #else and #endif.
 pp_if_indent_code               = false    # true/false
